Move BatchMessageListenerContainer to src/test
